Add ability for timer fields to support decimal seconds

+++ backend/gameplay_api/models.py
@@ -13,8 +13,8 @@ class ChessGame(models.Model):
 
 	halfmove_clock = models.IntegerField() # 50 move rule detection
 
-	white_player_clock = models.IntegerField() # In seconds
-	black_player_clock = models.IntegerField() # In seconds
+	white_player_clock = models.DecimalField() # In seconds
+	black_player_clock = models.DecimalField() # In seconds
 
 	captured_white_material = models.JSONField(default=dict)
 	captured_black_material = models.JSONField(default=dict)
